GYM-242: Add workoutBasePage to remove code duplication
